Retour arrière ACI 🔥

+++ client/components/arreteCadre/form/general.vue
@@ -1,18 +1,18 @@
 <script setup lang="ts">
-import { ArreteCadre } from '~/dto/arrete_cadre.dto';
-import { helpers, required } from '@vuelidate/validators/dist';
-import useVuelidate from '@vuelidate/core';
-import type { Ref } from 'vue';
-import type { Departement } from '~/dto/departement.dto';
-import { useAuthStore } from '~/stores/auth';
-import { useRefDataStore } from '~/stores/refData';
-import { requiredIf } from '@vuelidate/validators';
+import { ArreteCadre } from "~/dto/arrete_cadre.dto";
+import { helpers, required } from "@vuelidate/validators/dist";
+import useVuelidate from "@vuelidate/core";
+import type { Ref } from "vue";
+import type { Departement } from "~/dto/departement.dto";
+import { useAuthStore } from "~/stores/auth";
+import { useRefDataStore } from "~/stores/refData";
+import { requiredIf } from "@vuelidate/validators";
 
 const props = defineProps<{
   arreteCadre: ArreteCadre;
 }>();
 
-const query: Ref<string> = ref('');
+const query: Ref<string> = ref("");
 const departementsTags = ref([]);
 const departementsFiltered = ref([]);
 const utils = useUtils();
@@ -24,7 +24,7 @@ const assignDepartement = (force = false) => {
   if (!props.arreteCadre.id || force) {
     if (props.arreteCadre.departements.length < 1) {
       props.arreteCadre.departements =
-        authStore.user.role === 'departement' ? refDataStore.departements.filter((d) => d.code === authStore.user.roleDepartement) : [];
+        authStore.user.role === "departement" ? refDataStore.departements.filter((d) => d.code === authStore.user.roleDepartement) : [];
     } else if (props.arreteCadre.departements.length > 1 && !isAci.value) {
       props.arreteCadre.departements = [props.arreteCadre.departements[0]];
     }
@@ -41,34 +41,34 @@ onMounted(() => {
 const rules = computed(() => {
   return {
     numero: {
-      required: helpers.withMessage("Le numéro de l'arrêté est obligatoire", required),
+      required: helpers.withMessage("Le numéro de l'arrêté est obligatoire", required)
     },
     departements: {
-      required: helpers.withMessage("L'arrêté doit être lié à au moins un département", required),
-    },
+      required: helpers.withMessage("L'arrêté doit être lié à au moins un département", required)
+    }
   };
 });
 
 const filterDepartements = () => {
   departementsFiltered.value = query.value
     ? refDataStore.departements.filter((d) => {
-        return !props.arreteCadre.departements.map((ad) => ad.id).includes(d.id) && d.nom.toLowerCase().includes(query.value.toLowerCase());
-      })
+      return !props.arreteCadre.departements.map((ad) => ad.id).includes(d.id) && d.nom.toLowerCase().includes(query.value.toLowerCase());
+    })
     : [];
 };
 
 const departementsOptions = refDataStore.departements.map((d) => {
   return {
     value: d.id,
-    text: d.nom,
+    text: d.nom
   };
 });
 
 const selectDepartement = (departement: Departement) => {
-  if (typeof departement === 'string') {
+  if (typeof departement === "string") {
     return;
   }
-  query.value = '';
+  query.value = "";
   props.arreteCadre.departements = [...props.arreteCadre.departements, departement];
   computeDepartementsTags();
 };
@@ -81,16 +81,20 @@ const deleteDepartement = (departementId: number) => {
 const computeDepartementsTags = () => {
   departementsTags.value = props.arreteCadre.departements
     .filter((d, index) => {
-      return (authStore.user.role === 'mte' && index !== 0) || (authStore.user.role !== 'mte' && d.code !== authStore.user.roleDepartement);
+      return index !== 0;
     })
     .map((d) => {
       return {
         label: d.nom,
-        class: 'fr-tag--dismiss',
-        tagName: 'button',
+        class: (authStore.user.role !== "mte" && d.code === authStore.user.roleDepartement) ? "" :
+          "fr-tag--dismiss",
+        tagName: "button",
         onclick: () => {
+          if (authStore.user.role !== "mte" && d.code === authStore.user.roleDepartement) {
+            return;
+          }
           deleteDepartement(d.id);
-        },
+        }
       };
     });
 };
@@ -114,20 +118,20 @@ const v$ = useVuelidate(rules, props.arreteCadre);
 
 const aciOptions = [
   {
-    label: 'Interdépartemental',
-    value: true,
+    label: "Interdépartemental",
+    value: true
   },
   {
-    label: 'Départemental',
-    value: false,
-  },
+    label: "Départemental",
+    value: false
+  }
 ];
 
 watch(
   query,
   useUtils().debounce(async () => {
     filterDepartements();
-  }, 300),
+  }, 300)
 );
 
 watch(isAci, () => {
@@ -135,7 +139,7 @@ watch(isAci, () => {
 });
 
 defineExpose({
-  v$,
+  v$
 });
 </script>
